requirements.txt
fastapi>=0.115.8,<1.0
uvicorn[standard]>=0.34.0,<1.0
sqlalchemy>=2.0.38,<3.0
pydantic>=2.11.0,<3.0
python-dotenv>=1.0.1,<2.0
python-multipart>=0.0.20,<1.0
python-telegram-bot>=21.10,<22.0
